requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
openpyxl>=3.1.2
lxml>=5.0.0
icalendar>=5.0.0
pyremindkit @ git+https://github.com/namuan/pyremindkit@3e092e56c804d11339fcbd918ef647a9ef207f25
pyobjc-framework-EventKit>=10.0
